back-end/requirements.txt
pytube
pandas
matplotlib
google-auth
google-api-python-client
flask
flask-cors
pdfkit
secure-smtplib